refactor: embed all client logic component to client page
