+++ src/app/issues/[id]/author/layout.tsx
@@ -1,0 +1,13 @@
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+    title: "CommonGround - 撰寫觀點",
+};
+
+export default function RootLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return <>{children}</>;
+}
